move genEvent to it's own file
